requirements.txt
Flask
Flask-CORS
mysql-connector-python
bcrypt
opencv-python
face-recognition
numpy